adding some documentation
