+++ genepy/epigenetics/README.md
@@ -4,37 +4,32 @@ Especially targeted to functions related to the analysis of epigenomics data. It
 
 ## Available functions:
 
-_in ./chipseq.py_
+### chipseq.py
 
-- bigWigFrom
-- ReadRoseSuperEnhancers
-- loadPeaks
-- pysam_getPeaksAt
-- bedtools_getPeaksAt
-- makeProfiles
-- getPeaksAt
-- computeMeanCov
-- substractPeaks
-- simpleMergePeaks
-- findpeakpath
-- findBestPeak
-- mergeReplicatePeaks
-- col_nan_scatter
-- col_nan_kde_histo
-- findAdditionalPeaks
-- negLogLikelihood
-- poissonFit
-- putInBed
-- pairwiseOverlap
-- enrichment
-- findAdditionalCobindingSignal
-- fullDiffPeak
-- diffPeak
-- MakeSuperEnhancers
-- runChromHMM
-- loadMEMEmotifs
-- simpleMergeMotifs
-- substractPeaksTo
+- bigWigFrom: run the bigwig command line for a set of bam files in a folder
+- ReadRoseSuperEnhancers: reads ROSE2's output and returns its superenhancer bedfile as a pd dataframe. 
+- loadPeaks: loads 1 to many peak bedfile into one pandas dataframe.
+- simpleMergePeaks: simply merges bedfiles from peak callers. providing a concaneted dataframe of bed-like tables
+- putInBed: given a conscensus bed-like dataframe and another one, will merge the second one into the first
+- pairwiseOverlap: compute pairwise overlap and correlation on this overlap for a set of peaks mappe to a conscensus 
+- enrichment: compute pairwise enrichment and correlation for a set of peaks mappe to a conscensus 
+- fullDiffPeak: will use macs2 to call differential peak binding from two bam files and their control
+- diffPeak: calls MACS2 bdgdiff given some parameters
+- MakeSuperEnhancers: Calls super enhancer from H3K27ac with the ROSE algorithm
+- runChromHMM: runs the chromHMM algorithm
+- loadMEMEmotifs: loads motif from the output file of MEME after running fimo.
+- simpleMergeMotifs: aggregates the motifs if they overlap, into one motif file
+- substractPeaksTo: removes all peaks that are not within a bp distance to a set of loci
+
+### CREME.py
+
+The goal of creme is to be a simple, 1 function tool. It works with 1 to many sets of replicates for each pulled protein/mark.
+
+CREME takes as input a pandas dataframe. This dataframe is the concatenation of each replicates' bed files and can be loaded from a set of MACS2 bedfiles using genepy's loadPeaks function.
+
+CREME will output, amongst other thing, a dataframe representing a concatenation f bedfiles of merged replicates.
+
+find out more at __CREME.md__
 
 ## highly recommended packages
